Fix compiler warning
src/winscard_msg.c:489:17: warning: implicit conversion loses integer precision: 'ssize_t' (aka 'long') to 'int' [-Wshorten-64-to-32]
                        bytes_read = read(filedes, buffer, remaining);
                                   ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ src/winscard_msg.c
@@ -478,7 +478,7 @@ INTERNAL LONG MessageReceive(void *buffer_void, uint64_t buffer_size,
 		/* try to read only when socket is readable */
 		if (pollret > 0)
 		{
-			int bytes_read;
+			ssize_t bytes_read;
 
 			if (!(read_fd.revents & POLLIN))
 			{
